refactor : Update Security of UserDetail Logic
