--- src/main/java/com/flab/todo/common/config/security/PasswordEncoder.java
@@ -1,8 +1,0 @@
-package com.flab.todo.common.config.security;
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Component;
-
-@Component
-public class PasswordEncoder extends BCryptPasswordEncoder {
-}
