+++ src/main/java/com/flab/todo/common/config/security/SecurityConfig.java
@@ -6,38 +6,44 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.stereotype.Component;
 
 import com.flab.todo.member.MemberService;
 
 @EnableWebSecurity
 public class SecurityConfig {
-	private final MemberService memberService;
-	private final PasswordEncoder passwordEncoder;
+	private final UserDetailsService userDetailsService;
 	private final UnauthorizedEntryPoint unauthorizedEntryPoint;
 
-	public SecurityConfig(MemberService memberService, PasswordEncoder passwordEncoder,
+	public SecurityConfig(UserDetailsService userDetailsService,
 		UnauthorizedEntryPoint unauthorizedEntryPoint) {
-		this.memberService = memberService;
-		this.passwordEncoder = passwordEncoder;
+		this.userDetailsService = userDetailsService;
 		this.unauthorizedEntryPoint = unauthorizedEntryPoint;
 	}
 
 	@Bean
 	public CustomAuthenticationProvider customAuthenticationProvider() {
-		return new CustomAuthenticationProvider(memberService, passwordEncoder);
+		return new CustomAuthenticationProvider(userDetailsService, this.passwordEncoder());
 	}
 
 	@Bean
 	public AuthenticationManager authenticationManager(HttpSecurity http, BCryptPasswordEncoder bCryptPasswordEncoder,
 		MemberService memberService) throws Exception {
 		return http.getSharedObject(AuthenticationManagerBuilder.class)
-			.userDetailsService(memberService)
+			.userDetailsService(userDetailsService)
 			.passwordEncoder(bCryptPasswordEncoder)
 			.and()
 			.authenticationProvider(customAuthenticationProvider())
 			.build();
+	}
+
+	@Bean
+	public PasswordEncoder passwordEncoder(){
+		return new BCryptPasswordEncoder();
 	}
 
 	@Bean
