+++ src/main/java/com/flab/todo/common/config/security/user_detail/UserDetailServiceImpl.java
@@ -1,0 +1,24 @@
+package com.flab.todo.common.config.security.user_detail;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.stereotype.Service;
+
+import com.flab.todo.common.exception.custom.UnAuthorizedException;
+import com.flab.todo.database.entity.Member;
+import com.flab.todo.member.MemberMapper;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Service
+public class UserDetailServiceImpl implements UserDetailsService {
+
+	private final MemberMapper memberMapper;
+
+	@Override
+	public UserDetails loadUserByUsername(String username) {
+		Member findMember = memberMapper.findByEmail(username).orElseThrow(UnAuthorizedException::new);
+		return new UserDetailImpl(findMember);
+	}
+}
